# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tilecast API
  version: 0.10.0
  license:
    name: AGPL-3.0-only
paths:
...
  /api/v1/player/pairing-sessions:
    post:
      operationId: createPairingSession
      responses:
        "201": { description: Pairing session created }
        "409": { description: Installation identity mismatch }
        "429": { description: Pairing requests are rate limited }
  /api/v1/player/pairing-sessions/{id}:
    get:
      operationId: pollPairingSession
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            in: header,
            name: Authorization,
            required: true,
            schema: { type: string, pattern: "^Pairing " },
          }
      responses:
        "200": { description: Pairing state or one-time enrollment result }
        "401": { description: Poll secret missing or invalid }
  /api/v1/player/enroll:
    post:
      operationId: enrollPlayer
      responses:
        "201": { description: Device credential returned exactly once }
        "409": { description: Enrollment token already used }
  /api/v1/player/heartbeat:
    post:
      operationId: playerHeartbeat
      security: [{ deviceBearer: [] }]
      responses:
        "200":
          description:
            Heartbeat accepted. `data.ignoredFields` names optional playback
            identifiers that were malformed and therefore dropped.
        "400": { description: Heartbeat body rejected }
        "401": { description: Credential invalid or revoked }
        "403": { description: Screen disabled }
...
  /api/v1/screens:
    get:
      operationId: listScreens
      responses:
        "200": { description: Computed screen status list }
        "401": { description: Dashboard authentication required }
  /api/v1/screens/{id}:
    get:
      operationId: getScreen
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200": { description: Screen details }
    patch:
      operationId: updateScreen
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200": { description: Screen updated }
        "403": { description: Owner or Administrator required }
...
  /api/v1/screens/pairing/{id}/approve:
    post:
      operationId: approvePairing
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, roomName, roomNumber, description]
              properties:
                name: { type: string, minLength: 2, maxLength: 120 }
                locationId: { type: [string, "null"], format: uuid }
                roomName: { type: string, maxLength: 120 }
                roomNumber: { type: string, maxLength: 80 }
                description: { type: string, maxLength: 1000 }
                replaceExistingCredential:
                  type: boolean
                  default: false
                  description: Explicitly authorizes credential rotation after successful enrollment.
      responses:
        "200": { description: Screen approved }
        "409":
          {
            description: Pairing recovery confirmation required or pairing state conflict,
          }
...
  /api/v1/screens/pairing/{id}/reject:
    post:
      operationId: rejectPairing
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "204": { description: Pairing rejected }
  /api/v1/screens/{id}/disable:
    post:
      operationId: disableScreen
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses: { "204": { description: Screen disabled and disconnected } }
  /api/v1/screens/{id}/enable:
    post:
      operationId: enableScreen
      parameters:
        - {
            in: path,
            name: id,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses: { "204": { description: Screen enabled } }
...
components:
...
  securitySchemes:
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: Tilecast device credential
...